Contactos

+++ MyFirstApp/app/src/main/java/com/example/myfirstapp/HomeActivity.kt
@@ -1,6 +1,6 @@
 package com.example.myfirstapp
 
-import DrawPieChart
+import com.example.myfirstapp.ui.DrawPieChart
 import android.annotation.SuppressLint
 import android.content.Intent
 import android.os.Build
@@ -91,7 +91,7 @@ class HomeActivity : ComponentActivity() {
     }
 
     @Composable
-    private fun DrawSummary(gastosPorCategoria: Map<String, Double>) {
+    private fun DrawSummary(gastosPorCategoria: Map<String, Float>) {
         var indice = 0
         Column (
             modifier = Modifier
